+++ src/plugins/apollo/components/poll-picker-field/constants.js
@@ -1,4 +1,3 @@
 export default {
   CHANNEL_NAME: 'pollPicker',
-  REQUEST_COUNT: 5,
 };