+++ src/plugins/boost/components/sponsor-picker-field/constants.js
@@ -1,4 +1,3 @@
 export default {
   CHANNEL_NAME: 'sponsorPicker',
-  REQUEST_COUNT: 10,
 };