+++ src/plugins/ncr/components/node-picker-field/index.jsx
@@ -12,6 +12,7 @@ import React from 'react';
 import delegate from './delegate';
 import { selectActionTypes } from '../../constants';
 import selector from './selector';
+import './styles.scss';
 
 class NodePickerField extends React.Component {
   static propTypes = {
@@ -44,8 +45,9 @@ class NodePickerField extends React.Component {
     super(props);
 
     this.state = {
-      hasLoadedFirstSet: false,
       existingNodes: [],
+      hasLoadedFirstSet: false,
+      inputValue: '',
       menuListScrollTop: 0,
     };
 
@@ -54,7 +56,7 @@ class NodePickerField extends React.Component {
       trailing: true,
     });
     this.handleChange = this.handleChange.bind(this);
-    this.handleInputChange = debounce(this.handleInputChange.bind(this), 500);
+    this.handleInputChange = this.handleInputChange.bind(this);
     this.handleMenuOpen = this.handleMenuOpen.bind(this);
     this.handlePick = this.handlePick.bind(this);
     this.handleRemove = this.handleRemove.bind(this);
@@ -83,10 +85,11 @@ class NodePickerField extends React.Component {
   }
 
   componentWillUnmount() {
-    const { handleClearChannel } = this.props;
+    const { handleClearChannel, handleSearch } = this.props;
     handleClearChannel();
     this.handleLoadMore.cancel();
-    this.handleInputChange.cancel();
+    this.handleScroll.cancel();
+    handleSearch.cancel();
   }
 
   handleChange(selected, payload) {
@@ -108,10 +111,11 @@ class NodePickerField extends React.Component {
 
   handleInputChange(q) {
     const { handleSearch, isGetAll } = this.props;
-    if (isGetAll) {
-      return;
-    }
-    this.setState(() => ({ existingNodes: [] }), () => handleSearch(q));
+    this.setState(() => ({ inputValue: q }), () => {
+      if (!isGetAll) {
+        this.setState(() => ({ existingNodes: [] }), () => handleSearch(q));
+      }
+    });
   }
 
   handleLoadMore() {
@@ -119,12 +123,12 @@ class NodePickerField extends React.Component {
     if (isGetAll || !response.get('has_more')) {
       return;
     }
-    this.setState(({ existingNodes }, props) => ({
+    this.setState(({ existingNodes, menuListScrollTop }, props) => ({
       existingNodes: [
         ...existingNodes,
         ...props.response.get('nodes', []),
       ],
-      menuListScrollTop: this.menuList.scrollTop,
+      menuListScrollTop: get(this, 'menuList.scrollTop', menuListScrollTop),
     }), () => {
       handleSearch(response.get('ctx_request').get('q'), response.get('ctx_request').get('page') + 1);
     });
@@ -169,12 +173,13 @@ class NodePickerField extends React.Component {
       this.handleRemove(nodeRef);
     } else {
       this.handlePick(nodeRef);
+      this.setState(() => ({ inputValue: '' }));
     }
   }
 
   render() {
     const { fields, isMulti, selectComponents, response, value } = this.props;
-    const { existingNodes } = this.state;
+    const { existingNodes, inputValue } = this.state;
     const MenuComponent = selectComponents.Menu || components.Menu;
     const OptionComponent = selectComponents.Option || components.Option;
     return (
@@ -220,6 +225,7 @@ class NodePickerField extends React.Component {
         isMulti={isMulti}
         onChange={this.handleChange}
         onInputChange={this.handleInputChange}
+        inputValue={inputValue}
         onMenuOpen={this.handleMenuOpen}
         options={existingNodes.concat(!response ? [] : response.get('nodes', [])).map((node) => ({
           label: node.get('title'),