+++ src/plugins/ovp/components/video-picker-field/constants.js
@@ -1,4 +1,3 @@
 export default {
   CHANNEL_NAME: 'videoPicker',
-  REQUEST_COUNT: 5,
 };