+++ src/plugins/taxonomy/components/category-picker-field/constants.js
@@ -1,4 +1,3 @@
 export default {
   CHANNEL_NAME: 'categoryPicker',
-  REQUEST_COUNT: 10,
 };